requirements.txt
anthropic>=0.18.0
openai>=1.0.0
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
textblob>=0.17.1
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
